Consolidate Unit Tests
